+++ sshmitm/authentication.py
@@ -864,3 +864,25 @@ class AuthenticatorPassThrough(Authenticator):
             and self.pubkey_enumerator.connected
         ):
             self.pubkey_enumerator.close()
+
+
+class AuthenticatorRemote(Authenticator):
+
+    @classmethod
+    def parser_arguments(cls) -> None:
+        super().parser_arguments()
+        cls.argument_group()
+
+    def auth_publickey(self, username: str, host: str, port: int, key: PKey) -> int:
+        if key.can_sign():
+            logging.debug(
+                "AuthenticatorRemote.auth_publickey: username=%s, key=%s %s %sbits",
+                username,
+                key.get_name(),
+                key.fingerprint,
+                key.get_bits(),
+            )
+            return self.connect(
+                username, host, port, AuthenticationMethod.PUBLICKEY, key=key
+            )
+        return paramiko.common.AUTH_FAILED