+++ sshmitm/interfaces/server.py
@@ -441,7 +441,9 @@ class ServerInterface(BaseServerInterface):
         if not self.session.ssh_remote_channel:
             logging.error("session.ssh_remote_channel not initialized!")
             return False
-        self.session.ssh_remote_channel.resize_pty(width, height, pixelwidth, pixelheight)
+        self.session.ssh_remote_channel.resize_pty(
+            width, height, pixelwidth, pixelheight
+        )
         return True
 
     def check_channel_x11_request(  # pylint: disable=too-many-arguments